Moved vote code to vote-overview
