+++ app/routes/detail.ts
@@ -59,37 +59,8 @@ export default class DetailRoute extends Route {
       .sort(sortObjectsByTitle)
       .slice(0, 4);
 
-    const formattedTableVote: FormattedTableVote[] = [];
-
-    // add the votes to formattedTableVote in the form of [{proponent: {proponent1}, opponent: {oponent1}, abstainer: {abstainer1}},{proponent: {proponent2}, opponent: {oponent2}, abstainer: {abstainer2}},... ]
-    // iterate over all 3 arrays (proponent, opponent, abstainer) and push only one voter to the formattedTableVote array each iteration
-    // this means that the length of the formattedTableVote array will be the length of the longest array (proponent, opponent, abstainer)
-    // if there is no voter, push an empty object
-
     const agendaItemHandling = await agendaItem.handledBy;
     const vote = (await agendaItemHandling?.hasVotes)?.toArray().shift();
-
-    //TODO: The vote formatting is not the responssibility of the route. Move this logic to the component.
-    // https://binnenland.atlassian.net/browse/BNB-246
-    const proponents = (await vote?.hasProponents)?.toArray();
-    const opponents = (await vote?.hasOpponents)?.toArray();
-    const abstainers = (await vote?.hasAbstainers)?.toArray();
-    // get the length of the longest array (not the accumaleted length of all 3 arrays)
-    const longestArrayLength = Math.max(
-      proponents?.length || 0,
-      opponents?.length || 0,
-      abstainers?.length || 0
-    );
-
-    // iterate over the longest array
-    for (let i = 0; i < longestArrayLength; i++) {
-      // push the proponent, opponent and abstainer of the current iteration to the formattedTableVote array
-      formattedTableVote.push({
-        proponent: proponents?.[i] || null,
-        opponent: opponents?.[i] || null,
-        abstainer: abstainers?.[i] || null,
-      });
-    }
 
     const similiarAgendaItems = await this.store.query('agenda-item', {
       page: {
@@ -124,7 +95,6 @@ export default class DetailRoute extends Route {
       agendaItem,
       vote,
       agendaItemOnSameSession,
-      formattedTableVote,
       similiarAgendaItems,
     };
   }
